util: add get_uname_machine to detect userspace architecture
On ARM (e.g. Raspian OS) runs a 32bit userspace (e.g. armhf) on a Aarch64/ARM64
machine. Thus `uname -m` might be misleading as it identifies the kernel and
not the userspace. The introduced `get_uname_machine` checks the elf header of
the python interpreter itself to detect the architecture for dynamically linked
applications.

The terms align with the debian/ubuntu architecture naming (arm/armhf/aarch64/amd64).